added unsat benchmarks

+++ benchmarks/BV-benchamrks-unsat/bvisxdigit-32/isxdigit.h
@@ -6,6 +6,6 @@
 #include <assert.h>
 #include <inttypes.h>
 
-uint32_t isxdigit_(uint32_t c);
+uint8_t isxdigit_(uint8_t c);
 
 #endif
